Thought I'd do something if even only for practice (line algorithm)
The line drawer function. It compiles but I haven't run any tests on it,
and have no idea if the timings are off or anything. You didn't say to
do anything but I thought I might as well try something for practice.
Your line drawer looks a lot neater, but you seem to be getting both x0,
y0 and gray from the top level file or something. Thought I might as
well share this with you.